@@ -30,24 +30,24 @@
   },
   "devDependencies": {
     "@types/node": "^16.7.1",
-    "aws-cdk": "^1.119.0",
-    "standard-version": "^9.3.1",
+    "aws-cdk": "^2.35.0",
+    "standard-version": "^9.5.0",
     "ts-jest": "^27.0.5",
-    "ts-node": "^10.2.1",
-    "typescript": "^4.3.5"
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4"
   },
   "dependencies": {
-    "@aws-cdk/assert": "^1.119.0",
-    "@aws-cdk/aws-ssm": "^1.119.0",
-    "@aws-cdk/core": "^1.119.0",
+    "@aws-cdk/assert": "^2.35.0",
     "@types/flat": "5.0.2",
-    "@types/glob": "^7.1.4",
+    "@types/glob": "^7.2.0",
     "@types/jest": "^27.0.1",
     "@types/js-yaml": "^4.0.3",
+    "aws-cdk-lib": "^2.35.0",
+    "constructs": "^10.1.70",
     "flat": "^5.0.2",
-    "glob": "^7.1.7",
+    "glob": "^7.2.3",
     "jest": "^27.0.6",
     "js-yaml": "^4.1.0",
-    "source-map-support": "^0.5.19"
+    "source-map-support": "^0.5.21"
   }
 }
