@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
-import cdk = require('@aws-cdk/core');
+import * as cdk from 'aws-cdk-lib';
+
 import { ConfigExampleStack } from '../lib/config-example-stack';
 
 const app = new cdk.App();
